% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{accumulate}
\alias{accumulate}
\title{Accumulate items over trials, i.e. Guesses and Results.}
\usage{
accumulate(items, default = NA, ignore = NA)
}
\arguments{
\item{items}{The vector of items to accumulate in order.}

\item{default}{The default items. Optional. Defaults to NA.}

\item{ignore}{Item to ignore during accumulation. Optional.}
}
\description{
Accumulated items are lagged by one trial so that the resulting
item list represents the items that had been discovered up to
that trial.
}
\examples{
items <- c("a", "b", "b", "c")

accumulate(items)
# list(NA, c("a"), c("a", "b"), c("a", "b"))

accumulate(items, default = "a")
# list(c("a"), c("a"), c("a", "b"), c("a", "b"))

accumulate(items, ignore = "b")
# list(c("a"), c("a"), c("a"), c("a"))

}
