% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{accumulate_session}
\alias{accumulate_session}
\title{Accumulate guesses and inventory items by session.}
\usage{
accumulate_session(Guesses)
}
\description{
Accumulate guesses and inventory items by session.
}
