% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{assign_hashes}
\alias{assign_hashes}
\title{Assign a hash representing the vector of accumulated variables.}
\usage{
assign_hashes(accumulated)
}
\description{
Assign a hash representing the vector of accumulated variables.
}
