% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{assign_ids}
\alias{assign_ids}
\title{Assign a string id representing the vector of accumulated variables.}
\usage{
assign_ids(accumulated)
}
\description{
Assign a string id representing the vector of accumulated variables.
}
