% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-players.R
\name{assign_player_ix}
\alias{assign_player_ix}
\title{Assign PlayerIX, which enumerates players in teams}
\usage{
assign_player_ix(frame)
}
\description{
Assign PlayerIX, which enumerates players in teams
}
