% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots-data.R
\docType{data}
\name{BotsExplore}
\alias{BotsExplore}
\title{Results of simulations of robots playing the Totems game.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500000 rows and 18 columns.}
\usage{
data("BotsExplore")
}
\description{
In this experiment, each simulation is of a single player
playing this game for 5000 guesses. 100 indepdenent players
are simulated.
}
\keyword{datasets}
