% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventory-ids.R
\name{calculate_final_inventory_id}
\alias{calculate_final_inventory_id}
\title{Append a new result to an inventory.}
\usage{
calculate_final_inventory_id(prev_inventory_id, result)
}
\description{
Append a new result to an inventory.
}
