% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-performance.R
\name{clone_diachronic_teams}
\alias{clone_diachronic_teams}
\title{Duplicate guesses to allow summarizing diachronic team
performance at TeamSize == 2 and TeamSize == 4.}
\usage{
clone_diachronic_teams(diachronic)
}
\description{
Duplicate guesses to allow summarizing diachronic team
performance at TeamSize == 2 and TeamSize == 4.
}
