% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_db}
\alias{connect_db}
\title{Open a connection to the Totems MySQL DB.}
\usage{
connect_db()
}
\value{
DBI Connection class
}
\description{
Reads DB info from a file "config.yml"
expected in the current directory.
}
