% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{count_unique_combinations}
\alias{count_unique_combinations}
\title{Count unique combinations of items (with replacement).}
\usage{
count_unique_combinations(n_items)
}
\description{
Count unique combinations of items (with replacement).
}
