% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{count_unique_permutations}
\alias{count_unique_permutations}
\title{Count unique permutations (with replacement).}
\usage{
count_unique_permutations(n_items)
}
\description{
Count unique permutations (with replacement).
}
