% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_blank_config}
\alias{create_blank_config}
\title{Create a blank config file "config.yml" in the current directory.}
\usage{
create_blank_config()
}
\description{
Create a blank config file "config.yml" in the current directory.
}
