% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate.R
\name{cumulate_uniques}
\alias{cumulate_uniques}
\title{Cumulate unique items and guesses by player and by team}
\usage{
cumulate_uniques(frame)
}
\description{
Cumulate unique items and guesses by player and by team
}
