% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{data_guesses}
\alias{data_guesses}
\title{Get data for each guess made by players.}
\usage{
data_guesses()
}
\arguments{
\item{con}{DBI Connection class.}
}
\description{
Get data for each guess made by players.
}
