% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-players.R
\name{data_players}
\alias{data_players}
\title{Get player data from the database.}
\usage{
data_players(con)
}
\arguments{
\item{con}{DBIConnection class. See \code{\link{connect_db}}.}
}
\description{
Get player data from the database.
}
