% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teams.R
\name{data_teams}
\alias{data_teams}
\title{Get team data from the database.}
\usage{
data_teams(con)
}
\arguments{
\item{con}{DBIConnection class. See \code{\link{connect_db}}.}
}
\description{
Get team data from the database.
}
