% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{determine_current_players}
\alias{determine_current_players}
\title{Determine the number of current players}
\usage{
determine_current_players(frame)
}
\description{
Determine the number of current players
}
