% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{filter_50min}
\alias{filter_50min}
\title{Filter valid sessions and teams for the 50 minute experiment.}
\usage{
filter_50min(frame = NULL)
}
\description{
Filter valid sessions and teams for the 50 minute experiment.
}
