% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{filter_instructions}
\alias{filter_instructions}
\title{Filter valid sessions and teams for the inherited instructions experiment.}
\usage{
filter_instructions(frame = NULL)
}
\description{
Filter valid sessions and teams for the inherited instructions experiment.
}
