% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{filter_selfother}
\alias{filter_selfother}
\title{Filter valid sessions and teams for the self-other experiment.}
\usage{
filter_selfother(frame = NULL)
}
\description{
Filter valid sessions and teams for the self-other experiment.
}
