% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{filter_teamsize}
\alias{filter_teamsize}
\title{Filter valid sessions and teams for the scalability experiment.}
\usage{
filter_teamsize(frame = NULL)
}
\description{
Filter valid sessions and teams for the scalability experiment.
}
