% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-guess.R
\name{get_item_filename}
\alias{get_item_filename}
\title{Get the image filename for an item from it's image number.}
\usage{
get_item_filename(item_number, item_labels)
}
\description{
Get the image filename for an item from it's image number.
}
