% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-guess.R
\name{get_item_image}
\alias{get_item_image}
\title{Get the path to an item image based on its number.}
\usage{
get_item_image(item_number, item_labels)
}
\description{
Get the path to an item image based on its number.
}
