% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{Guesses}
\alias{Guesses}
\title{Workshop guesses made by players in the Totems experiment.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41732 rows and 13 columns.}
\usage{
data("Guesses")
}
\description{
The raw observations for the Totems experiment are individual
workshop guesses, consisting of a Guess (e.g., "0_0_1_3") and a
Result, which is 0 if the guess was incorrect, and the number
of the created item if correct.
}
\details{
\describe{
  \item{PlayerID}{unique identifier for players}
  \item{PlayerTime}{seconds, time elapsed for this player when this guess was made}
  \item{TeamTime}{seconds, time elapsed for this team when this guess was made}
  \item{GuessNum}{current guess number}
  \item{TeamGuessNum}{current team guess number}
  \item{Guess}{a string created from a combination of up to 4 items that represents a guess}
  \item{Result}{numeric, the item number of the result. Result == 0 if the guess is incorrect.}
  \item{Score}{numeric, the score, if any, received for this guess. Normal innovations are 15 points each. Anything that is used to make a Totem is worth much more than any tool. More advanced totems are worth more points.}
  \item{UniqueGuess}{bool, has this guess been made by this player before?}
  \item{TeamUniqueGuess}{bool, has this guess been made by this team before?}
  \item{UniqueItem}{bool, has this item been made by this player before?}
  \item{TeamUniqueItem}{bool, has this item been made by this team before?}
}
}
\keyword{datasets}
