% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inheritance.R
\name{highlight_inheritance_100}
\alias{highlight_inheritance_100}
\title{Create a new column Inheritance that labels G4 Diachronic players
in the 100 labor minute experiment.}
\usage{
highlight_inheritance_100(frame)
}
\description{
Inheritance is "diachronic_inheritance" for the G4 Diachronic player,
and "no_inheritance" for all other session types.
}
