% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{Inventories}
\alias{Inventories}
\title{Inventories uncovered by teams in the Totems experiment.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 979 rows and 9 columns.}
\usage{
data("Inventories")
}
\description{
\describe{
  \item{TeamID}{unique identifier for teams}
  \item{TeamInventory}{a string representing all items that have been discovered by this team}
  \item{NumInnovations}{number of new items discovered by the team}
  \item{NumAdjacent}{number of discoveries that could be made from this inventory}
  \item{TeamGuesses}{number of guesses it took to make a new discovery}
  \item{TeamUniqueGuesses}{number of unique guesses for the team it took to make a new discovery}
  \item{UniqueGuesses}{number of unique guesses for the players it took to make a new discovery}
  \item{Duration}{the amount of time it took the team to make a new discovery}
  \item{Difficulty}{the difficulty of this inventory, taken as a function of
    the number of items that can be combined relative to the number of adajcent
    and correct items}
}
}
\keyword{datasets}
