% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-experiment.R
\name{label_experiment}
\alias{label_experiment}
\title{Label the experiment of the observations in frame.}
\usage{
label_experiment(frame)
}
\arguments{
\item{frame}{Requires columns TeamID and PlayerID.}
}
\value{
The same data as in frame, but with new columns Exp and TeamSize
}
\description{
Data for Gen1 and Gen2 Diachronic players is
duplicated: once for the 50 labor minutes comparison,
and once for the 100 labor minutes comparison.
}
