% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players-and-sessions.R
\name{label_generation}
\alias{label_generation}
\title{Label Generation based on Strategy, PlayerIX, and SessionIX.}
\usage{
label_generation(frame)
}
\description{
Label Generation based on Strategy, PlayerIX, and SessionIX.
}
