% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{label_players_per_session}
\alias{label_players_per_session}
\title{Label the number of players in a session based on Strategy and Size.}
\usage{
label_players_per_session(frame)
}
\description{
For Synchronic players, PlayersPerSession == Size.
For Diachronic and Isolated players, PlayersPerSession == 1.
}
