% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{label_result_uniqueness}
\alias{label_result_uniqueness}
\title{Label results as unique and accumulate them.}
\usage{
label_result_uniqueness(Guesses)
}
\description{
Label results as unique and accumulate them.
}
