% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{label_session_duration}
\alias{label_session_duration}
\title{Label the duration of the session for players by TeamID.}
\usage{
label_session_duration(frame)
}
\description{
Label the duration of the session for players by TeamID.
}
