% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players-and-sessions.R
\name{label_session_player}
\alias{label_session_player}
\title{Label PlayerID, PlayerIX, and SessionIX from SessionID}
\usage{
label_session_player(frame)
}
\description{
Label PlayerID, PlayerIX, and SessionIX from SessionID
}
