% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{label_stage}
\alias{label_stage}
\title{Label the stage of the guesser's inventory relative to the team inventory.}
\usage{
label_stage(Guesses)
}
\description{
Label the stage of the guesser's inventory relative to the team inventory.
}
