% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage.R
\name{label_stage_ix}
\alias{label_stage_ix}
\title{Label stage ix for individual players inheriting either
from self or another teammate.}
\usage{
label_stage_ix(IndividualGuesses)
}
\description{
Label stage ix for individual players inheriting either
from self or another teammate.
}
