% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{label_team_score}
\alias{label_team_score}
\title{Score guesses that resulted in unique team items}
\usage{
label_team_score(frame, scores)
}
\description{
Score guesses that resulted in unique team items
}
