% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{label_unique_guesses}
\alias{label_unique_guesses}
\title{Assess the uniqueness of each Guess for this player.}
\usage{
label_unique_guesses(frame)
}
\description{
Assess the uniqueness of each Guess for this player.
}
