% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{label_unique_items}
\alias{label_unique_items}
\title{Assess the uniqueness of each created Item for this player.}
\usage{
label_unique_items(Guesses)
}
\description{
Assess the uniqueness of each created Item for this player.
}
