% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{label_unique_session_results}
\alias{label_unique_session_results}
\title{Assess the uniqueness of results created in this session.}
\usage{
label_unique_session_results(frame)
}
\description{
Assess the uniqueness of results created in this session.
}
