% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{label_unique_team_guesses}
\alias{label_unique_team_guesses}
\title{Assess the uniqueness of each guess for this team.}
\usage{
label_unique_team_guesses(frame)
}
\description{
Assess the uniqueness of each guess for this team.
}
