% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{label_unique_team_items}
\alias{label_unique_team_items}
\title{Assess the uniqueness of each created Item for this team.}
\usage{
label_unique_team_items(frame)
}
\description{
Assess the uniqueness of each created Item for this team.
}
