% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{label_unique_team_results}
\alias{label_unique_team_results}
\title{Assess the uniqueness of each created Item for this team.}
\usage{
label_unique_team_results(frame)
}
\description{
Assess the uniqueness of each created Item for this team.
}
