% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{load_totems_theme}
\alias{load_totems_theme}
\title{Return a list containing theme elements to be used across visualizations.}
\usage{
load_totems_theme()
}
\description{
Return a list containing theme elements to be used across visualizations.
}
