% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{PlayerInfo}
\alias{PlayerInfo}
\title{Information about players}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 177 rows and 6 columns.}
\usage{
data("PlayerInfo")
}
\description{
\describe{
  \item{PlayerID}{unique identifier for players}
  \item{TeamID}{unique identifier for teams}
  \item{Strategy}{strategy assigned to players in this team}
  \item{Generation}{generation of participant. Only relevant for Diachronic teams.}
  \item{SessionDuration}{the length of time this player played the game in a single session}
  \item{PlayerIX}{index for each player in a team, e.g., P1 and P2. Assigned at random.}
}
}
\keyword{datasets}
