% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_csvs}
\alias{read_csvs}
\title{Util function for loading a directory of csvs
and assigning the resulting data frames with
prefixed names.}
\usage{
read_csvs(directory, prefix = "", stem_func = function(x) x)
}
\description{
Util function for loading a directory of csvs
and assigning the resulting data frames with
prefixed names.
}
