% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_attempt_measures}
\alias{recode_attempt_measures}
\title{Recode the labels of the type of attempt measure.}
\usage{
recode_attempt_measures(frame)
}
\description{
Recode the labels of the type of attempt measure.
}
