% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{recode_generation_type}
\alias{recode_generation_type}
\title{Recode generation type for 50 labor minute teams.}
\usage{
recode_generation_type(frame)
}
\description{
Recode generation type for 50 labor minute teams.
}
