% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teams.R
\name{recode_group_id}
\alias{recode_group_id}
\title{Recode ID_Group as TeamID without datetime information.}
\usage{
recode_group_id(frame, con)
}
\description{
Recode ID_Group as TeamID without datetime information.
}
