% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess-type.R
\name{recode_guess_type}
\alias{recode_guess_type}
\title{Recode guess type based on the uniqueness of the guess, whether the guess
created an item, and the uniqueness of the created item.}
\usage{
recode_guess_type(frame, unique_guess, unique_result, suffix)
}
\arguments{
\item{unique_guess}{str Name of column in frame containing uniqueness of
the guess, e.g., "UniqueSessionGuess"}

\item{unique_result}{str Name of column in frame containing uniqueness of
result, e.g., "UniqueSessionResult"}
}
\value{
Same as input with new columns GuessType and GuessTypeLabel.
}
\description{
The reason for the params `unique_guess` and `unique_result` is that
uniqueness can be determined relative to either the session or to the team.
}
