% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_score_value}
\alias{recode_score_value}
\title{Recode the labels of the type of scoring metric.}
\usage{
recode_score_value(frame)
}
\description{
Recode the labels of the type of scoring metric.
}
