% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{recode_session_type_100}
\alias{recode_session_type_100}
\title{Create a session type column that compares individual session types
for the 100 labor minute experiment.}
\usage{
recode_session_type_100(frame)
}
\description{
Session types are:
- Diachronic G1
- Diachronic G2
- Diachronic G3
- Diachronic G4
- Isolated S1
- Isolated S2
- Isolated S3
- Isolated S4
- Synchronic 4
}
