% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{recode_session_type_50min}
\alias{recode_session_type_50min}
\title{Create a session type column that compares individual session types
for the 50 labor minute experiment.}
\usage{
recode_session_type_50min(frame)
}
\description{
Session types are:
- Diachronic G1
- Diachronic G2
- Isolated 50min
- Synchronic 2
}
