% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_strategy}
\alias{recode_strategy}
\title{Recode strategy condition by merging with a map of labels.}
\usage{
recode_strategy(frame)
}
\description{
If no frame is provided, just the map of labels is returned.
}
