% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players-and-sessions.R
\name{replace_ancestor}
\alias{replace_ancestor}
\title{Create Generation from Strategy and Ancestor.}
\usage{
replace_ancestor(frame)
}
\description{
For Synchronic teams, Generation = 1.
For Diachronic and Isolated teams, Generation = Ancestor.
}
