% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teams.R
\name{replace_id_group}
\alias{replace_id_group}
\title{Replace ID_Group with TeamID.}
\usage{
replace_id_group(frame)
}
\description{
This function deidentifies ID_Group by removing
datetime information from it.
}
