% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{replace_trial_time}
\alias{replace_trial_time}
\title{Replace TrialTime (msec) with PlayerTime (sec)}
\usage{
replace_trial_time(frame)
}
\description{
Replace TrialTime (msec) with PlayerTime (sec)
}
