% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{report_glm_mod}
\alias{report_glm_mod}
\title{Report the results of a glm model.}
\usage{
report_glm_mod(glm_mod, term, min_p_value = 0.001, p_value_only = FALSE)
}
\description{
Report the results of a glm model.
}
