% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{report_lm_mod}
\alias{report_lm_mod}
\title{Report the results of an lm model.}
\usage{
report_lm_mod(lm_mod, term, min_p_value = 0.001, p_value_only = FALSE)
}
\description{
Report the results of an lm model.
}
