% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-utils.R
\name{report_lmer_mod}
\alias{report_lmer_mod}
\title{Report the results of an lmer model.}
\usage{
report_lmer_mod(lmer_mod, term, formats = NULL, reverse_sign = FALSE)
}
\description{
Report the results of an lmer model.
}
