% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{rolling}
\alias{rolling}
\title{Keep track of accumulating variables.}
\usage{
rolling(items, default = NA)
}
\description{
Keep track of accumulating variables.
}
\examples{
items <- c("a", "b", "b", "c")
rolling(items)

}
