% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{rolling_player_guesses}
\alias{rolling_player_guesses}
\title{Create PrevGuesses list column containing
unique guesses made by this player.}
\usage{
rolling_player_guesses(Guesses)
}
\description{
Create PrevGuesses list column containing
unique guesses made by this player.
}
