% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{rolling_player_inventory}
\alias{rolling_player_inventory}
\title{Create PrevInventory list column containing
unique inventories held by this player.}
\usage{
rolling_player_inventory(Guesses)
}
\description{
Create PrevInventory list column containing
unique inventories held by this player.
}
