% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{rolling_team_guesses}
\alias{rolling_team_guesses}
\title{Create PrevTeamGuesses list column containing
unique guesses made by this team.}
\usage{
rolling_team_guesses(Guesses)
}
\description{
Create PrevTeamGuesses list column containing
unique guesses made by this team.
}
