% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-guesses.R
\name{rolling_team_inventory}
\alias{rolling_team_inventory}
\title{Create PrevTeamInventory list column containing
unique inventories held by this team.}
\usage{
rolling_team_inventory(Guesses)
}
\description{
Create PrevTeamInventory list column containing
unique inventories held by this team.
}
