% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guesses.R
\name{sample_session}
\alias{sample_session}
\title{Sample guesses at regular intervals in a session.}
\usage{
sample_session(session_guesses, default, interval = 1)
}
\description{
Requires SessionID, SessionDuration, SessionTime, NumSessionGuesses,
PrevSessionInventoryID, SessionInventorySize,
PrevSessionGuessesHash, NumUniqueSessionGuesses
}
