% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Sampled}
\alias{Sampled}
\title{Cumulative performance sampled at a regular interval.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38603 rows and 20 columns.}
\usage{
data("Sampled")
}
\description{
Cumulative performance sampled at a regular interval.
}
\keyword{datasets}
