% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{SampledPerformance}
\alias{SampledPerformance}
\title{Cumulative performance sampled at a regular interval.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5766 rows and 9 columns.}
\usage{
data("SampledPerformance")
}
\description{
Cumulative performance sampled at a regular interval.
}
\keyword{datasets}
