% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-drive.R
\name{save_survey_responses}
\alias{save_survey_responses}
\title{Save survey responses from Google Drive spreadsheet.}
\usage{
save_survey_responses()
}
\description{
Save survey responses from Google Drive spreadsheet.
}
