% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Sessions}
\alias{Sessions}
\title{Information about player sessions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 834 rows and 8 columns.}
\usage{
data("Sessions")
}
\description{
Information about player sessions
}
\keyword{datasets}
