% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{summarize_team_performance}
\alias{summarize_team_performance}
\title{Summarize the team's final performance.}
\usage{
summarize_team_performance(frame)
}
\description{
Summarize the team's final performance.
}
