% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totems-data.R
\docType{data}
\name{TeamInfo}
\alias{TeamInfo}
\title{Information about teams}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 96 rows and 3 columns.}
\usage{
data("TeamInfo")
}
\description{
Information about teams
}
\keyword{datasets}
