% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Teams}
\alias{Teams}
\title{Information about teams.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 457 rows and 4 columns.}
\usage{
data("Teams")
}
\description{
Information about teams.
}
\keyword{datasets}
