% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-guess.R
\name{view_guess}
\alias{view_guess}
\title{View a guess as pictures of items added together.}
\usage{
view_guess(guess_str, item_labels, view = FALSE)
}
\description{
View a guess as pictures of items added together.
}
