% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the Totems MySQL DB.}
\usage{
connect_db(config_file = "config.yml")
}
\arguments{
\item{config_file}{Config file in yaml format. Defaults to "config.yml".}
}
\description{
Connect to the Totems MySQL DB.
}
