% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_blank_config}
\alias{create_blank_config}
\title{Create a blank config file.}
\usage{
create_blank_config(config_file = "config.yml")
}
\arguments{
\item{config_file}{Name of config file in yaml format. Defaults to "config.yml".}
}
\description{
Create a blank config file.
}
