% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Guesses}
\alias{Guesses}
\title{Workshop guesses made by players in the Totems experiment.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 169173 rows and 34 columns.}
\usage{
data("Guesses")
}
\description{
The raw observations for the Totems experiment are individual
workshop guesses, consisting of a Guess (e.g., "0_0_1_3") and a
Result, which is 0 if the guess was incorrect, and the number
of the created item if correct.
}
\keyword{datasets}
