% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{replace_id_group}
\alias{replace_id_group}
\title{Replace ID_Group with TeamID.}
\usage{
replace_id_group(frame)
}
\description{
TeamID is a deidentified version of ID_Group
with datetime information removed.
}
