% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players-and-sessions.R
\name{replace_id_player}
\alias{replace_id_player}
\title{Replace ID_Player with PlayerID, PlayerIX, SessionID, and SessionIX.}
\usage{
replace_id_player(frame)
}
\description{
For all players:
SessionID = "S" + ID_Player
}
\details{
For Diachronic and Synchronic players:
PlayerID = "P" + ID_Player
PlayerIX = 1:4?
SessionIX = 1

For Isolated players, who are assigned multiple values for ID_Player:
PlayerID = "P" + min(ID_Player)
PlayerIX = 1
SessionIX = 1:4?
}
